Heroku-Binance/requirements.txt
pandas_ta==0.2.45b0
pandas==1.2.4
pause==0.3
tabulate==0.8.9
pushbullet.py==0.12.0
python_binance==1.0.10
binance==0.3
